moved flask routes to their own file routes.py. this is to prepare for client>>server communication for creating a habit
